inlined the moderation widget, changed default export, and defined typed consts in withtracker
